requirements.txt
pandas ~= 2.1.1
azure-storage-blob ~= 12.21.0
aiohttp ~= 3.9.5
python-dotenv ~= 1.0.1
openpyxl ~= 3.1.5
build == 1.2.1
xlrd ~= 2.0.1
scipy ~= 1.14.0
tqdm ~= 4.66.1
numpy == 1.26.4
asyncpg == 0.29.0
click ~= 8.1.7
pydantic ~= 2.8.2
papermill ~= 2.6.0
notebook ~= 7.2.1
country-converter ~= 1.2